# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SwiftRoute Logistics
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuoteRequestServiceType =
  (typeof QuoteRequestServiceType)[keyof typeof QuoteRequestServiceType];

export const QuoteRequestServiceType = {
  standard: "standard",
  express: "express",
  freight: "freight",
  international: "international",
} as const;

export interface QuoteRequest {
  name: string;
  email: string;
  phone?: string;
  origin: string;
  destination: string;
  weight?: number;
  serviceType: QuoteRequestServiceType;
  notes?: string;
}

export interface QuoteResponse {
  id: number;
  name: string;
  email: string;
  phone?: string;
  origin: string;
  destination: string;
  weight?: number;
  serviceType: string;
  notes?: string;
  status: string;
  createdAt: string;
}

export interface TrackingEvent {
  timestamp: string;
  location: string;
  description: string;
  status: string;
}

export interface TrackingResult {
  trackingNumber: string;
  status: string;
  origin: string;
  destination: string;
  estimatedDelivery: string;
  events: TrackingEvent[];
}

export interface Service {
  id: number;
  name: string;
  description: string;
  icon: string;
  features: string[];
  estimatedDays: string;
  startingPrice: number;
}

export interface Stats {
  shipmentsDelivered: number;
  countriesServed: number;
  clientsServed: number;
  onTimeDeliveryRate: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}