# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SwiftRoute Logistics
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: quotes
    description: Shipment quote requests
  - name: tracking
    description: Shipment tracking
  - name: services
    description: Logistics services
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /quotes:
    post:
      operationId: submitQuote
      tags: [quotes]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "201":
          description: Quote submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quote requests
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteResponse"

  /tracking/{trackingNumber}:
    get:
      operationId: trackShipment
      tags: [tracking]
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
      summary: Track a shipment
      responses:
        "200":
          description: Shipment tracking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingResult"
        "404":
          description: Shipment not found

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all logistics services
      responses:
        "200":
          description: List of logistics services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /stats:
    get:
      operationId: getStats
      tags: [services]
      summary: Get company stats summary
      responses:
        "200":
          description: Company statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    QuoteRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        origin:
          type: string
        destination:
          type: string
        weight:
          type: number
        serviceType:
          type: string
          enum: [standard, express, freight, international]
        notes:
          type: string
      required:
        - name
        - email
        - origin
        - destination
        - serviceType

    QuoteResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        origin:
          type: string
        destination:
          type: string
        weight:
          type: number
        serviceType:
          type: string
        notes:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - origin
        - destination
        - serviceType
        - status
        - createdAt

    TrackingResult:
      type: object
      properties:
        trackingNumber:
          type: string
        status:
          type: string
        origin:
          type: string
        destination:
          type: string
        estimatedDelivery:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
      required:
        - trackingNumber
        - status
        - origin
        - destination
        - estimatedDelivery
        - events

    TrackingEvent:
      type: object
      properties:
        timestamp:
          type: string
        location:
          type: string
        description:
          type: string
        status:
          type: string
      required:
        - timestamp
        - location
        - description
        - status

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: array
          items:
            type: string
        estimatedDays:
          type: string
        startingPrice:
          type: number
      required:
        - id
        - name
        - description
        - icon
        - features
        - estimatedDays
        - startingPrice

    Stats:
      type: object
      properties:
        shipmentsDelivered:
          type: integer
        countriesServed:
          type: integer
        clientsServed:
          type: integer
        onTimeDeliveryRate:
          type: number
      required:
        - shipmentsDelivered
        - countriesServed
        - clientsServed
        - onTimeDeliveryRate

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error