# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SwiftRoute Logistics
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a quote request
 */
export const SubmitQuoteBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().optional(),
  origin: zod.string(),
  destination: zod.string(),
  weight: zod.number().optional(),
  serviceType: zod.enum(["standard", "express", "freight", "international"]),
  notes: zod.string().optional(),
});

/**
 * @summary List all quote requests
 */
export const ListQuotesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().optional(),
  origin: zod.string(),
  destination: zod.string(),
  weight: zod.number().optional(),
  serviceType: zod.string(),
  notes: zod.string().optional(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListQuotesResponse = zod.array(ListQuotesResponseItem);

/**
 * @summary Track a shipment
 */
export const TrackShipmentParams = zod.object({
  trackingNumber: zod.coerce.string(),
});

export const TrackShipmentResponse = zod.object({
  trackingNumber: zod.string(),
  status: zod.string(),
  origin: zod.string(),
  destination: zod.string(),
  estimatedDelivery: zod.string(),
  events: zod.array(
    zod.object({
      timestamp: zod.string(),
      location: zod.string(),
      description: zod.string(),
      status: zod.string(),
    }),
  ),
});

/**
 * @summary List all logistics services
 */
export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  icon: zod.string(),
  features: zod.array(zod.string()),
  estimatedDays: zod.string(),
  startingPrice: zod.number(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Get company stats summary
 */
export const GetStatsResponse = zod.object({
  shipmentsDelivered: zod.number(),
  countriesServed: zod.number(),
  clientsServed: zod.number(),
  onTimeDeliveryRate: zod.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});